+++ src/main/java/com/mycompany/webapp/dto/ClassReview.java
@@ -15,5 +15,5 @@ public class ClassReview { //클래스 리뷰
 	private String mid; //->회원 아이디
 	
 	private String mnickname;
-	private Integer avgCrratio; //별점 평균(정수로 반올림)
+	private Float avgCrratio; //별점 평균(정수로 반올림)
 }
